+++ CHANGES.md
@@ -7,6 +7,7 @@ History
 * Add `autoClose` option to `launch` fucntion
 * Add `loop` option to `launch` function (experimental)
 * Add `Page.setBypassCSP` method
+* Fix `slowMo` option
 * Continue to find WS URL while process is alive
 
 ## Version 0.0.20 (2018-08-11)